Add random bytes implementation
