+++ .travis.yml
@@ -4,6 +4,7 @@ php:
     - 5.5
     - 5.6
     - 7.0
+    - 7.1
 
 branches:
     only:
@@ -18,6 +19,8 @@ matrix:
         - php: 5.6
           env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
         - php: 7.0
+          env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
+        - php: 7.1
           env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
 
 before_script:
