+++ composer.json
@@ -17,9 +17,9 @@
         "symfony/security-core": "~2.4||~3.0"
     },
     "require-dev": {
-        "knplabs/phpspec-welldone-extension": "dev-master@dev",
+        "bossa/phpspec2-expect": "~1.0",
         "pedrotroller/php-cs-custom-fixer": "~1.5.0",
-        "phpspec/phpspec": "~2.4"
+        "phpspec/phpspec": "~2.5"
     },
     "autoload": {
         "psr-4": {
@@ -32,7 +32,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.1.x-dev"
+            "dev-master": "2.2.x-dev"
         }
     }
 }
