+++ phpspec.yml
@@ -1,4 +1,1 @@
 formatter.name: pretty
-
-extensions:
-    - Knp\PhpSpec\WellDone\Extension
