+++ src/Knp/Rad/User/HasPassword/HasPassword.php
@@ -8,11 +8,19 @@ trait HasPassword
 {
     use HasInitialPassword;
 
+    /**
+     * @return string
+     */
     public function getPassword()
     {
         return $this->password;
     }
 
+    /**
+     * @param string $password
+     *
+     * @return self
+     */
     public function setPassword($password)
     {
         $this->password = $password;
