+++ src/Knp/Rad/User/HasSalt.php
@@ -2,6 +2,9 @@
 
 namespace Knp\Rad\User;
 
+/**
+ * @deprecated The salt feature is deprecated since PHP 5.5 and BCrypt usage. Please upgrade your version of PHP and use BCrypt
+ */
 interface HasSalt
 {
     /**
