+++ src/Knp/Rad/User/HasSalt/HasSalt.php
@@ -4,11 +4,19 @@ namespace Knp\Rad\User\HasSalt;
 
 trait HasSalt
 {
+    /**
+     * @return string
+     */
     public function getSalt()
     {
         return $this->salt;
     }
 
+    /**
+     * @param string $salt
+     *
+     * @return self
+     */
     public function setSalt($salt)
     {
         $this->salt = $salt;
