+++ src/Knp/Rad/User/Resources/config/services.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+  <parameters>
+    <parameter key="knp_rad_user.salt.generator.random_bytes.length">255</parameter>
+  </parameters>
+
   <services>
 
     <service id="knp_rad_user.event_listener.odm.password_hash_listener" class="Knp\Rad\User\EventListener\ODM\PasswordHashListener" public="false">
@@ -26,7 +30,14 @@
       <tag name="doctrine.event_listener" event="prePersist" method="prePersist" priority="200"/>
       <tag name="doctrine_mongodb.odm.event_listener" event="prePersist" method="prePersist" priority="200"/>
       <tag name="doctrine_couchdb.event_listener" event="prePersist" method="prePersist" priority="200"/>
+      <argument type="service" id="knp_rad_user.salt.generator"/>
     </service>
+
+    <service id="knp_rad_user.salt.generator.pseudo_random_bytes_generator" class="Knp\Rad\User\Salt\Generator\PseudoRandomBytesGenerator" public="false" lazy="true">
+      <argument type="parameter" id="knp_rad_user.salt.generator.random_bytes.length"/>
+    </service>
+
+    <service id="knp_rad_user.salt.generator" alias="knp_rad_user.salt.generator.pseudo_random_bytes_generator" />
 
   </services>
 </container>
